Pulled out more steppers.
This time: BitSet, FlatHashTable, HashTable, and immutable sets and maps
(trie-based).
