+++ src/main/scala/scala/compat/java8/StepConverters.scala
@@ -32,313 +32,6 @@ package converterImpls {
     def nextLong() = if (proxied ne null) proxied.nextStep else underlying.next
   }
 
-  private[java8] class StepsAnyFlatHashTable[A](_underlying: Array[AnyRef], _i0: Int, _iN: Int)
-  extends StepsLikeGapped[A, StepsAnyFlatHashTable[A]](_underlying, _i0, _iN) {
-    def next() = if (currentEntry eq null) throwNSEE else { val ans = currentEntry.asInstanceOf[A]; currentEntry = null; ans }
-    def semiclone(half: Int) = new StepsAnyFlatHashTable[A](underlying, i0, half)
-  }
-
-  private[java8] class StepsDoubleFlatHashTable(_underlying: Array[AnyRef], _i0: Int, _iN: Int)
-  extends StepsDoubleLikeGapped[StepsDoubleFlatHashTable](_underlying, _i0, _iN) {
-    def nextDouble() = if (currentEntry eq null) throwNSEE else { val ans = currentEntry.asInstanceOf[Double]; currentEntry = null; ans }
-    def semiclone(half: Int) = new StepsDoubleFlatHashTable(underlying, i0, half)    
-  }
-
-  private[java8] class StepsIntFlatHashTable(_underlying: Array[AnyRef], _i0: Int, _iN: Int)
-  extends StepsIntLikeGapped[StepsIntFlatHashTable](_underlying, _i0, _iN) {
-    def nextInt() = if (currentEntry eq null) throwNSEE else { val ans = currentEntry.asInstanceOf[Int]; currentEntry = null; ans }
-    def semiclone(half: Int) = new StepsIntFlatHashTable(underlying, i0, half)    
-  }
-
-  private[java8] class StepsLongFlatHashTable(_underlying: Array[AnyRef], _i0: Int, _iN: Int)
-  extends StepsLongLikeGapped[StepsLongFlatHashTable](_underlying, _i0, _iN) {
-    def nextLong() = if (currentEntry eq null) throwNSEE else { val ans = currentEntry.asInstanceOf[Long]; currentEntry = null; ans }
-    def semiclone(half: Int) = new StepsLongFlatHashTable(underlying, i0, half)    
-  }
-
-  private[java8] class StepsAnyHashTableKey[K, HE <: collection.mutable.HashEntry[K, HE]](_underlying: Array[collection.mutable.HashEntry[K, HE]], _i0: Int, _iN: Int)
-  extends StepsLikeGapped[K, StepsAnyHashTableKey[K, HE]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def next() = if (currentEntry eq null) throwNSEE else { val ans = currentEntry.asInstanceOf[HE].key; currentEntry = currentEntry.asInstanceOf[HE].next; ans }
-    def semiclone(half: Int) = new StepsAnyHashTableKey[K, HE](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, HE]]], i0, half)
-  }
-
-  private[java8] class StepsDoubleHashTableKey[HE <: collection.mutable.HashEntry[Double, HE]](_underlying: Array[collection.mutable.HashEntry[Double, HE]], _i0: Int, _iN: Int)
-  extends StepsDoubleLikeGapped[StepsDoubleHashTableKey[HE]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def nextDouble() = if (currentEntry eq null) throwNSEE else { val ans = currentEntry.asInstanceOf[HE].key; currentEntry = currentEntry.asInstanceOf[HE].next; ans }
-    def semiclone(half: Int) = new StepsDoubleHashTableKey[HE](underlying.asInstanceOf[Array[collection.mutable.HashEntry[Double, HE]]], i0, half)
-  }
-
-  private[java8] class StepsIntHashTableKey[HE <: collection.mutable.HashEntry[Int, HE]](_underlying: Array[collection.mutable.HashEntry[Int, HE]], _i0: Int, _iN: Int)
-  extends StepsIntLikeGapped[StepsIntHashTableKey[HE]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def nextInt() = if (currentEntry eq null) throwNSEE else { val ans = currentEntry.asInstanceOf[HE].key; currentEntry = currentEntry.asInstanceOf[HE].next; ans }
-    def semiclone(half: Int) = new StepsIntHashTableKey[HE](underlying.asInstanceOf[Array[collection.mutable.HashEntry[Int, HE]]], i0, half)
-  }
-
-  private[java8] class StepsLongHashTableKey[HE <: collection.mutable.HashEntry[Long, HE]](_underlying: Array[collection.mutable.HashEntry[Long, HE]], _i0: Int, _iN: Int)
-  extends StepsLongLikeGapped[StepsLongHashTableKey[HE]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def nextLong() = if (currentEntry eq null) throwNSEE else { val ans = currentEntry.asInstanceOf[HE].key; currentEntry = currentEntry.asInstanceOf[HE].next; ans }
-    def semiclone(half: Int) = new StepsLongHashTableKey[HE](underlying.asInstanceOf[Array[collection.mutable.HashEntry[Long, HE]]], i0, half)
-  }
-
-  private[java8] class StepsAnyDefaultHashTable[K, V](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, V]]], _i0: Int, _iN: Int)
-  extends StepsLikeGapped[(K, V), StepsAnyDefaultHashTable[K, V]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def next() = 
-      if (currentEntry eq null) throwNSEE
-      else { val de = currentEntry.asInstanceOf[collection.mutable.DefaultEntry[K, V]]; currentEntry = de.next; (de.key, de.value) }
-    def semiclone(half: Int) =
-      new StepsAnyDefaultHashTable[K, V](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, V]]]], i0, half)
-  }
-
-  private[java8] class StepsAnyDefaultHashTableValue[K, V](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, V]]], _i0: Int, _iN: Int)
-  extends StepsLikeGapped[V, StepsAnyDefaultHashTableValue[K, V]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def next() = 
-      if (currentEntry eq null) throwNSEE
-      else { val de = currentEntry.asInstanceOf[collection.mutable.DefaultEntry[K, V]]; currentEntry = de.next; de.value }
-    def semiclone(half: Int) =
-      new StepsAnyDefaultHashTableValue[K, V](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, V]]]], i0, half)
-  }
-
-  private[java8] class StepsDoubleDefaultHashTableValue[K](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, Double]]], _i0: Int, _iN: Int)
-  extends StepsDoubleLikeGapped[StepsDoubleDefaultHashTableValue[K]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def nextDouble() = 
-      if (currentEntry eq null) throwNSEE
-      else { val de = currentEntry.asInstanceOf[collection.mutable.DefaultEntry[K, Double]]; currentEntry = de.next; de.value }
-    def semiclone(half: Int) =
-      new StepsDoubleDefaultHashTableValue[K](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, Double]]]], i0, half)
-  }
-
-  private[java8] class StepsIntDefaultHashTableValue[K](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, Int]]], _i0: Int, _iN: Int)
-  extends StepsIntLikeGapped[StepsIntDefaultHashTableValue[K]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def nextInt() = 
-      if (currentEntry eq null) throwNSEE
-      else { val de = currentEntry.asInstanceOf[collection.mutable.DefaultEntry[K, Int]]; currentEntry = de.next; de.value }
-    def semiclone(half: Int) =
-      new StepsIntDefaultHashTableValue[K](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, Int]]]], i0, half)
-  }
-
-  private[java8] class StepsLongDefaultHashTableValue[K](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, Long]]], _i0: Int, _iN: Int)
-  extends StepsLongLikeGapped[StepsLongDefaultHashTableValue[K]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def nextLong() = 
-      if (currentEntry eq null) throwNSEE
-      else { val de = currentEntry.asInstanceOf[collection.mutable.DefaultEntry[K, Long]]; currentEntry = de.next; de.value }
-    def semiclone(half: Int) =
-      new StepsLongDefaultHashTableValue[K](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.DefaultEntry[K, Long]]]], i0, half)
-  }
-
-  private[java8] class StepsAnyLinkedHashTable[K, V](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, V]]], _i0: Int, _iN: Int)
-  extends StepsLikeGapped[(K, V), StepsAnyLinkedHashTable[K, V]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def next() = 
-      if (currentEntry eq null) throwNSEE
-      else { val de = currentEntry.asInstanceOf[collection.mutable.LinkedEntry[K, V]]; currentEntry = de.next; (de.key, de.value) }
-    def semiclone(half: Int) =
-      new StepsAnyLinkedHashTable[K, V](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, V]]]], i0, half)
-  }
-
-  private[java8] class StepsAnyLinkedHashTableValue[K, V](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, V]]], _i0: Int, _iN: Int)
-  extends StepsLikeGapped[V, StepsAnyLinkedHashTableValue[K, V]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def next() = 
-      if (currentEntry eq null) throwNSEE
-      else { val de = currentEntry.asInstanceOf[collection.mutable.LinkedEntry[K, V]]; currentEntry = de.next; de.value }
-    def semiclone(half: Int) =
-      new StepsAnyLinkedHashTableValue[K, V](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, V]]]], i0, half)
-  }
-
-  private[java8] class StepsDoubleLinkedHashTableValue[K](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, Double]]], _i0: Int, _iN: Int)
-  extends StepsDoubleLikeGapped[StepsDoubleLinkedHashTableValue[K]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def nextDouble() = 
-      if (currentEntry eq null) throwNSEE
-      else { val de = currentEntry.asInstanceOf[collection.mutable.LinkedEntry[K, Double]]; currentEntry = de.next; de.value }
-    def semiclone(half: Int) =
-      new StepsDoubleLinkedHashTableValue[K](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, Double]]]], i0, half)
-  }
-
-  private[java8] class StepsIntLinkedHashTableValue[K](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, Int]]], _i0: Int, _iN: Int)
-  extends StepsIntLikeGapped[StepsIntLinkedHashTableValue[K]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def nextInt() = 
-      if (currentEntry eq null) throwNSEE
-      else { val de = currentEntry.asInstanceOf[collection.mutable.LinkedEntry[K, Int]]; currentEntry = de.next; de.value }
-    def semiclone(half: Int) =
-      new StepsIntLinkedHashTableValue[K](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, Int]]]], i0, half)
-  }
-
-  private[java8] class StepsLongLinkedHashTableValue[K](_underlying: Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, Long]]], _i0: Int, _iN: Int)
-  extends StepsLongLikeGapped[StepsLongLinkedHashTableValue[K]](_underlying.asInstanceOf[Array[AnyRef]], _i0, _iN) {
-    def nextLong() = 
-      if (currentEntry eq null) throwNSEE
-      else { val de = currentEntry.asInstanceOf[collection.mutable.LinkedEntry[K, Long]]; currentEntry = de.next; de.value }
-    def semiclone(half: Int) =
-      new StepsLongLinkedHashTableValue[K](underlying.asInstanceOf[Array[collection.mutable.HashEntry[K, collection.mutable.LinkedEntry[K, Long]]]], i0, half)
-  }
-
-  private[java8] class StepsAnyImmHashMap[K, V](_underlying: collection.immutable.HashMap[K, V], _i0: Int, _iN: Int)
-  extends StepsLikeImmHashMap[K, V, (K, V), StepsAnyImmHashMap[K, V]](_underlying, _i0, _iN) {
-    protected def demiclone(u: collection.immutable.HashMap[K,V], j0: Int, jN: Int) = new StepsAnyImmHashMap[K, V](u, j0, jN)
-    def next(): (K, V) = 
-      if (hasNext) { 
-        if (theIterator eq null) theIterator = underlying.iterator
-        i += 1
-        theIterator.next
-      }
-      else throwNSEE  
-  }
-
-  private[java8] class StepsAnyImmHashMapKey[K, V](_underlying: collection.immutable.HashMap[K, V], _i0: Int, _iN: Int)
-  extends StepsLikeImmHashMap[K, V, K, StepsAnyImmHashMapKey[K, V]](_underlying, _i0, _iN) {
-    protected def demiclone(u: collection.immutable.HashMap[K,V], j0: Int, jN: Int) = new StepsAnyImmHashMapKey[K, V](u, j0, jN)
-    def next(): K = 
-      if (hasNext) { 
-        if (theIterator eq null) theIterator = underlying.keysIterator
-        i += 1
-        theIterator.next
-      }
-      else throwNSEE  
-  }
-
-  private[java8] class StepsAnyImmHashMapValue[K, V](_underlying: collection.immutable.HashMap[K, V], _i0: Int, _iN: Int)
-  extends StepsLikeImmHashMap[K, V, V, StepsAnyImmHashMapValue[K, V]](_underlying, _i0, _iN) {
-    protected def demiclone(u: collection.immutable.HashMap[K,V], j0: Int, jN: Int) = new StepsAnyImmHashMapValue[K, V](u, j0, jN)
-    def next(): V = 
-      if (hasNext) { 
-        if (theIterator eq null) theIterator = underlying.valuesIterator
-        i += 1
-        theIterator.next
-      }
-      else throwNSEE  
-  }
-
-  private[java8] class StepsDoubleImmHashMapKey[V](_underlying: collection.immutable.HashMap[Double, V], _i0: Int, _iN: Int)
-  extends StepsDoubleLikeImmHashMap[Double, V, StepsDoubleImmHashMapKey[V]](_underlying, _i0, _iN) {
-    protected def demiclone(u: collection.immutable.HashMap[Double,V], j0: Int, jN: Int) = new StepsDoubleImmHashMapKey[V](u, j0, jN)
-    def nextDouble() = 
-      if (hasNext) { 
-        if (theIterator eq null) theIterator = underlying.keysIterator
-        i += 1
-        theIterator.next
-      }
-      else throwNSEE  
-  }
-
-  private[java8] class StepsDoubleImmHashMapValue[K](_underlying: collection.immutable.HashMap[K, Double], _i0: Int, _iN: Int)
-  extends StepsDoubleLikeImmHashMap[K, Double, StepsDoubleImmHashMapValue[K]](_underlying, _i0, _iN) {
-    protected def demiclone(u: collection.immutable.HashMap[K,Double], j0: Int, jN: Int) = new StepsDoubleImmHashMapValue[K](u, j0, jN)
-    def nextDouble() = 
-      if (hasNext) { 
-        if (theIterator eq null) theIterator = underlying.valuesIterator
-        i += 1
-        theIterator.next
-      }
-      else throwNSEE  
-  }
-
-  private[java8] class StepsIntImmHashMapKey[V](_underlying: collection.immutable.HashMap[Int, V], _i0: Int, _iN: Int)
-  extends StepsIntLikeImmHashMap[Int, V, StepsIntImmHashMapKey[V]](_underlying, _i0, _iN) {
-    protected def demiclone(u: collection.immutable.HashMap[Int,V], j0: Int, jN: Int) = new StepsIntImmHashMapKey[V](u, j0, jN)
-    def nextInt() = 
-      if (hasNext) { 
-        if (theIterator eq null) theIterator = underlying.keysIterator
-        i += 1
-        theIterator.next
-      }
-      else throwNSEE  
-  }
-
-  private[java8] class StepsIntImmHashMapValue[K](_underlying: collection.immutable.HashMap[K, Int], _i0: Int, _iN: Int)
-  extends StepsIntLikeImmHashMap[K, Int, StepsIntImmHashMapValue[K]](_underlying, _i0, _iN) {
-    protected def demiclone(u: collection.immutable.HashMap[K,Int], j0: Int, jN: Int) = new StepsIntImmHashMapValue[K](u, j0, jN)
-    def nextInt() = 
-      if (hasNext) { 
-        if (theIterator eq null) theIterator = underlying.valuesIterator
-        i += 1
-        theIterator.next
-      }
-      else throwNSEE  
-  }
-
-  private[java8] class StepsLongImmHashMapKey[V](_underlying: collection.immutable.HashMap[Long, V], _i0: Int, _iN: Int)
-  extends StepsLongLikeImmHashMap[Long, V, StepsLongImmHashMapKey[V]](_underlying, _i0, _iN) {
-    protected def demiclone(u: collection.immutable.HashMap[Long,V], j0: Int, jN: Int) = new StepsLongImmHashMapKey[V](u, j0, jN)
-    def nextLong() = 
-      if (hasNext) { 
-        if (theIterator eq null) theIterator = underlying.keysIterator
-        i += 1
-        theIterator.next
-      }
-      else throwNSEE  
-  }
-
-  private[java8] class StepsLongImmHashMapValue[K](_underlying: collection.immutable.HashMap[K, Long], _i0: Int, _iN: Int)
-  extends StepsLongLikeImmHashMap[K, Long, StepsLongImmHashMapValue[K]](_underlying, _i0, _iN) {
-    protected def demiclone(u: collection.immutable.HashMap[K,Long], j0: Int, jN: Int) = new StepsLongImmHashMapValue[K](u, j0, jN)
-    def nextLong() = 
-      if (hasNext) { 
-        if (theIterator eq null) theIterator = underlying.valuesIterator
-        i += 1
-        theIterator.next
-      }
-      else throwNSEE  
-  }
-
-  private[java8] class StepsAnyImmHashSet[A](_underlying: Iterator[A], _N: Int)
-  extends StepsLikeTrieIterator[A, StepsAnyImmHashSet[A]](_underlying, _N) {
-    protected def demiclone(it: Iterator[A], N: Int) = new StepsAnyImmHashSet(it, N)
-    def next(): A = { val ans = underlying.next; i += 1; ans }
-  }
-
-  private[java8] class StepsDoubleImmHashSet(_underlying: Iterator[Double], _N: Int)
-  extends StepsDoubleLikeTrieIterator[StepsDoubleImmHashSet](_underlying, _N) {
-    protected def demiclone(it: Iterator[Double], N: Int) = new StepsDoubleImmHashSet(it, N)
-    def nextDouble() = { val ans = underlying.next; i += 1; ans }
-  }
-
-  private[java8] class StepsIntImmHashSet(_underlying: Iterator[Int], _N: Int)
-  extends StepsIntLikeTrieIterator[StepsIntImmHashSet](_underlying, _N) {
-    protected def demiclone(it: Iterator[Int], N: Int) = new StepsIntImmHashSet(it, N)
-    def nextInt() = { val ans = underlying.next; i += 1; ans }
-  }
-
-  private[java8] class StepsLongImmHashSet(_underlying: Iterator[Long], _N: Int)
-  extends StepsLongLikeTrieIterator[StepsLongImmHashSet](_underlying, _N) {
-    protected def demiclone(it: Iterator[Long], N: Int) = new StepsLongImmHashSet(it, N)
-    def nextLong() = { val ans = underlying.next; i += 1; ans }
-  }
-
-  private[java8] class StepsIntBitSet(_underlying: Array[Long], _i0: Int, _iN: Int)
-  extends StepsIntLikeSliced[Array[Long], StepsIntBitSet](_underlying, _i0, _iN) {
-    private var mask: Long = (-1L) << (i & 0x3F)
-    private var cache: Long = underlying(i >>> 6)
-    private var found: Boolean = false
-    def semiclone(half: Int) = {
-      val ans = new StepsIntBitSet(underlying, i, half)
-      i = half
-      mask = (-1L) << (i & 0x3F)
-      cache = underlying(i >>> 6)
-      found = false
-      ans
-    }
-    def hasNext(): Boolean = found || ((i < iN) && {
-      while ((mask & cache) == 0) {
-        i += java.lang.Long.numberOfLeadingZeros(~mask)
-        if (i < 0 || i >= iN) { i = iN; return false }
-        mask = -1L
-        cache = underlying(i >>> 6)
-      }
-      var m = mask << 1
-      while ((mask & cache) == (m & cache)) {
-        mask = m
-        m = mask << 1
-        i += 1
-      }
-      if (i < 0 || i >= iN) {
-        i = iN
-        false
-      }
-      else {
-        found = true
-        true
-      }
-    })
-    def nextInt() = if (hasNext) { val j = i; found = false; mask = mask << 1; i += 1; j } else throwNSEE
-  }
 
   final class RichArrayAnyCanStep[A](private val underlying: Array[A]) extends AnyVal with MakesAnyStepper[A] {
     @inline def stepper: AnyStepper[A] with EfficientSubstep = new StepsAnyArray[A](underlying, 0, underlying.length)
